requirements.txt
numpy
pandas
matplotlib
seaborn
scikit-learn
sentence-transformers
faiss-cpu
pytest
fastapi
uvicorn
streamlit